requirements.txt
streamlit
langchain
faiss-cpu
transformers
openpyxl
pandas
plotly
openai
google-genai
groq
python-dotenv
sentence-transformers
langchain-community
langchain-openai
langchain-google-genai
langchain-groq
pypdf
langchain-huggingface
accelerate
Pillow
langgraph